pyproject.toml
[project]
name = "astro-immo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "cloudinary>=1.44.2",
    "fastapi>=0.136.0",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.45.0",
]
